split paths tab
